+++ modules/serverpod_auth/serverpod_auth_shared_flutter/lib/src/image_uploader.dart
@@ -1,15 +1,19 @@
 import 'dart:typed_data';
 
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:image/image.dart' as img;
-import 'package:image_picker/image_picker.dart';
-import 'package:file_picker/file_picker.dart';
 import 'package:image_cropper/image_cropper.dart';
+import 'package:image_picker/image_picker.dart';
 
 import 'session_manager.dart';
 
+/// Use the image uploader to upload avatars associated with a users account.
 class ImageUploader {
+  /// Uploads a new image for a user. The user will be asked to pick a file,
+  /// and is offered to crop the image (cropping is not available on web).
+  /// The file is then uploaded to the server. An
   static Future<void> updateUserImage({
     required BuildContext context,
     required SessionManager sessionManager,
@@ -26,7 +30,6 @@ class ImageUploader {
       var image = img.decodeImage(result.files.first.bytes!);
       if (image == null) return;
 
-      print('loaded image');
       if (image.width > image.height) {
         var h = imageSize;
         var w = imageSize * image.width / image.height;
@@ -45,8 +48,6 @@ class ImageUploader {
 
       image = img.copyResizeCropSquare(image, imageSize);
 
-      print('resized image');
-
       // Encode as png
       var encoded = img.encodePng(image);
       var bytes = Uint8List.fromList(encoded);
@@ -59,7 +60,7 @@ class ImageUploader {
       var toolbarWidgetColor =
           Theme.of(context).buttonTheme.colorScheme?.onPrimary ?? Colors.white;
 
-      final picker = ImagePicker();
+      var picker = ImagePicker();
       // Pick an image
       var imageFile = await picker.pickImage(source: ImageSource.gallery);
       if (imageFile == null) return;
@@ -69,7 +70,7 @@ class ImageUploader {
         sourcePath: imageFile.path,
         maxHeight: imageSize,
         maxWidth: imageSize,
-        aspectRatio: CropAspectRatio(ratioX: 1.0, ratioY: 1.0),
+        aspectRatio: const CropAspectRatio(ratioX: 1.0, ratioY: 1.0),
         cropStyle: CropStyle.circle,
         androidUiSettings: AndroidUiSettings(
             toolbarTitle: 'Cropper',
@@ -77,7 +78,7 @@ class ImageUploader {
             toolbarWidgetColor: toolbarWidgetColor,
             initAspectRatio: CropAspectRatioPreset.original,
             lockAspectRatio: false),
-        iosUiSettings: IOSUiSettings(
+        iosUiSettings: const IOSUiSettings(
           minimumAspectRatio: 1.0,
         ),
       );
@@ -88,8 +89,9 @@ class ImageUploader {
       var image = img.decodeImage(await croppedImageFile.readAsBytes());
       if (image == null) return;
 
-      if (image.width != imageSize || image.height != imageSize) ;
-      image = img.copyResize(image, width: imageSize, height: imageSize);
+      if (image.width != imageSize || image.height != imageSize) {
+        image = img.copyResize(image, width: imageSize, height: imageSize);
+      }
 
       // Encode as png
       var encoded = img.encodePng(image);