+++ modules/serverpod_auth/serverpod_auth_shared_flutter/lib/src/loading_barrier.dart
@@ -4,7 +4,7 @@ class _LoadingBarrier extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      child: Center(
+      child: const Center(
         child: CircularProgressIndicator(),
       ),
     );