+++ modules/serverpod_auth/serverpod_auth_server/CHANGELOG.md
@@ -1,3 +1,50 @@
-## 1.0.0
+## 0.9.2
 
-- Initial version, created by Stagehand
+- Adds serverpod_auth module for authentication with email, Apple, and Google.
+
+## 0.9.1
+
+- Fixes broken images in documentation.
+
+## 0.9.0
+
+- Updates documentation and logos
+- Ready for 0.9 release!
+
+## 0.8.12
+
+- Updates default templates.
+
+## 0.8.11
+
+- Improved ORM.
+- Support for Docker.
+- Chat module.
+- Updated docs.
+
+## 0.8.10
+
+- Support for static file directories in Relic.
+- Adds logos (psd and pngs).
+- Adds example project.
+- Initial version of authentication module.
+- Cloud storage support.
+- Adds auth module
+
+## 0.8.6
+
+- Adds documentation.
+- Generates SQL files for creating database tables.
+
+## 0.8.5
+
+- Fixes compilation in broken serverpod_cli package
+
+## 0.8.4
+
+- Updates template files and fixes `serverpod create` command.
+- Adds CHANGELOG.md
+
+## 0.8.3
+
+- Initial working version.